TST: move test that requires interactive backend to subprocess
